requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
opencv-python
pymysql
scipy
matplotlib
